Move atob and btoa to buffer
